refactor failure checker
the new checking logic does not use string contains
which makes it more durable if there is output which
happens to match the expected failure text.

new logic checks succesful results against all policies
which have been run, and logs a failure if any policies
did not yield a success.